Support fractional shares
- models.py: change shares column from Integer to Numeric(12,4) in
  Portfolio and Transaction
- application.py: parse shares as float instead of int in buy and sell;
  use a small epsilon (0.0001) when checking if all shares are sold to
  avoid float equality issues
- buy.html / sell.html: add step="any" and lower min to allow decimals

Requires a schema migration in Neon:
  ALTER TABLE portfolio ALTER COLUMN shares TYPE NUMERIC(12,4);
  ALTER TABLE histories ALTER COLUMN shares TYPE NUMERIC(12,4);

+++ models.py
@@ -22,7 +22,7 @@ class Portfolio(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     user_id = db.Column(db.Integer, db.ForeignKey("users.id"), nullable=False)
     symbol = db.Column(db.String(10), nullable=False)
-    shares = db.Column(db.Integer, nullable=False)
+    shares = db.Column(db.Numeric(12, 4), nullable=False)
     price = db.Column(db.Numeric(12, 2), nullable=False)
     total = db.Column(db.Numeric(12, 2), nullable=False)
 
@@ -33,6 +33,6 @@ class Transaction(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     user_id = db.Column(db.Integer, db.ForeignKey("users.id"), nullable=False)
     symbol = db.Column(db.String(10), nullable=False)
-    shares = db.Column(db.Integer, nullable=False)
+    shares = db.Column(db.Numeric(12, 4), nullable=False)
     price = db.Column(db.Numeric(12, 2), nullable=False)
     datetime = db.Column(db.DateTime, default=lambda: datetime.now(timezone.utc))
